pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flux2-klein-lora"
version = "0.1.0"
description = "LoRA trainer for FLUX.2-klein-4B - Fast, memory-efficient fine-tuning"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Limbicnation", email = "gero@limbicnation.com"},
]
keywords = ["flux", "lora", "training", "diffusers", "pixel-art"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision",
    "diffusers>=0.32.0",
    "transformers>=4.44.0",
    "accelerate>=0.34.0",
    "peft>=0.13.0",
    "huggingface-hub>=0.25.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "tqdm",
    "safetensors",
    "omegaconf",
    "wandb",
    "click>=8.0.0",
]

[project.scripts]
flux2-train = "flux2_klein_trainer.cli:main"
klein-lora = "flux2_klein_trainer.cli:main"

[project.urls]
Homepage = "https://github.com/Limbicnation/flux2-klein-lora"
Repository = "https://github.com/Limbicnation/flux2-klein-lora"
Documentation = "https://github.com/Limbicnation/flux2-klein-lora#readme"

[tool.hatch.build.targets.wheel]
packages = ["flux2_klein_trainer"]
